# docs/contracts/http/commands.openapi.yaml
openapi: 3.1.0
info:
  title: HomeTusk Commands API
  version: 1.0.0
  description: |
    Intent-driven API for household task management.
    Users submit commands, not CRUD operations.

    **Stage 1:** Structured commands (no NL processing yet).

    See also:
    - [CLAUDE.md](../../../CLAUDE.md) — Development rules
    - [ADR-003](../../architecture/decisions/003-stage1-commands-api.md) — Stage 1 decisions
servers:
  - url: /api/v1
    description: API version 1
security:
  - bearerAuth: []
...
paths:
  /commands:
    post:
      operationId: executeCommand
      summary: Execute a household command
      description: |
        Processes a command and returns the execution result.
        Commands are first-class entities with their own lifecycle.

        **Supported command types:**
        - `create_task` — Create a new task
        - `complete_task` — Mark a task as done

        **Traceability:**
        - Provide `X-Correlation-ID` header for distributed tracing
        - Server generates one if not provided
        - All logs and database records include this ID
      tags:
        - Commands
      parameters:
        - name: X-Correlation-ID
          in: header
          required: false
          description: Client-provided correlation ID for distributed tracing
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: false
          description: |
            Idempotency key for safe retries. Unique per user for 24 hours.
            Reuse with a different payload returns 409 IDEMPOTENCY_CONFLICT.
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: '^[A-Za-z0-9._-]{1,128}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandRequest'
            examples:
              createTask:
                summary: Create a task
                value:
                  householdId: "123e4567-e89b-12d3-a456-426614174000"
                  type: "create_task"
                  payload:
                    title: "Clean the kitchen"
                    zoneId: "123e4567-e89b-12d3-a456-426614174001"
                    deadline: "2026-01-13T18:00:00Z"
                  source: "web"
              completeTask:
                summary: Complete a task
                value:
                  householdId: "123e4567-e89b-12d3-a456-426614174000"
                  type: "complete_task"
                  payload:
                    taskId: "123e4567-e89b-12d3-a456-426614174002"
                  source: "mobile"
      responses:
        '200':
          description: Command processed (status in response body)
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID for this request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResponse'
        '400':
          description: Invalid request (schema/validation/type errors)
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: User does not have access to this household
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
        '409':
          description: Idempotency conflict (key reuse with different payload or in-progress)
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
  /commands/{commandId}/continue:
    post:
      operationId: continueCommand
      summary: Continue a command awaiting input
      tags:
        - Commands
      parameters:
        - name: commandId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Correlation-ID
          in: header
          required: false
          description: Client-provided correlation ID for distributed tracing
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContinueCommandRequest'
      responses:
        '200':
          description: Command processed (status in response body)
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
              description: Correlation ID for this request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResponse'
        '400':
          description: Command cannot be continued in current state
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Access denied for this command
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
        '404':
          description: Command not found
          headers:
            X-Correlation-ID:
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
  /users/me:
    get:
      operationId: getCurrentUser
      summary: Get current user profile
      description: |
        Returns the current user's profile.
        If the user does not exist, creates a new profile from JWT claims.
      tags:
        - Users
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak JWT token
  schemas:
...
    CommandRequest:
      type: object
      required:
        - householdId
        - type
        - payload
        - source
      properties:
        householdId:
          type: string
          format: uuid
          description: Target household for this command
        type:
          type: string
          enum:
            - create_task
            - complete_task
          description: Command type to execute
        payload:
          oneOf:
            - $ref: '#/components/schemas/CreateTaskPayload'
            - $ref: '#/components/schemas/CompleteTaskPayload'
          description: Command-specific payload
        source:
          type: string
          enum:
            - api
            - web
            - mobile
          description: Origin of the command
        clientTimestamp:
          type: string
          format: date-time
          description: Client-side timestamp (ISO 8601)
    ContinueCommandRequest:
      type: object
      required:
        - additionalInput
      properties:
        additionalInput:
          type: object
          additionalProperties: true
          description: Additional input to merge with original command payload
    CreateTaskPayload:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title
        description:
          type: string
          maxLength: 2000
          description: Task description
        zoneId:
          type: string
          format: uuid
          description: Zone/location for the task (must exist in household)
        assigneeId:
          type: string
          format: uuid
          description: Explicit assignee (must be household member)
        deadline:
          type: string
          format: date-time
          description: Deadline (must be in the future)
    CompleteTaskPayload:
      type: object
      required:
        - taskId
      properties:
        taskId:
          type: string
          format: uuid
          description: ID of task to complete (must exist in household)
    CommandResponse:
      oneOf:
        - $ref: '#/components/schemas/CommandExecutedResponse'
        - $ref: '#/components/schemas/CommandNeedsInputResponse'
        - $ref: '#/components/schemas/CommandRejectedResponse'
        - $ref: '#/components/schemas/CommandDegradedResponse'
      discriminator:
        propertyName: status
        mapping:
          executed: '#/components/schemas/CommandExecutedResponse'
          needs_input: '#/components/schemas/CommandNeedsInputResponse'
          rejected: '#/components/schemas/CommandRejectedResponse'
          executed_degraded: '#/components/schemas/CommandDegradedResponse'
    CommandExecutedResponse:
      type: object
      required:
        - commandId
        - correlationId
        - status
        - result
        - executionMs
        - initiatorId
      properties:
        commandId:
          type: string
          format: uuid
          description: Unique ID of the created command
        correlationId:
          type: string
          format: uuid
          description: Correlation ID for tracing
        status:
          type: string
          enum:
            - executed
          description: Command executed successfully
        result:
          $ref: '#/components/schemas/CommandResult'
        executionMs:
          type: integer
          minimum: 0
          description: Server-side execution time in milliseconds
        initiatorId:
          type: string
          format: uuid
          description: User who initiated this command (from JWT)
    CommandNeedsInputResponse:
      type: object
      required:
        - commandId
        - correlationId
        - status
        - question
        - requiredFields
        - executionMs
        - initiatorId
      properties:
        commandId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - needs_input
        question:
          type: string
          description: Question to ask the user
        requiredFields:
          type: array
          items:
            type: string
          description: Fields that need user input
        suggestions:
          type: object
          additionalProperties: true
          description: Suggested values for fields
        policyName:
          type: string
          description: Guardrail policy that requested clarification
        executionMs:
          type: integer
          minimum: 0
        initiatorId:
          type: string
          format: uuid
    CommandRejectedResponse:
      type: object
      required:
        - commandId
        - correlationId
        - status
        - errorCode
        - reason
        - executionMs
        - initiatorId
      properties:
        commandId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - rejected
        errorCode:
          type: string
          description: Machine-readable rejection code
        reason:
          type: string
          description: Human-readable rejection reason
        executionMs:
          type: integer
          minimum: 0
        initiatorId:
          type: string
          format: uuid
    CommandDegradedResponse:
      type: object
      required:
        - commandId
        - correlationId
        - status
        - result
        - executionMs
        - initiatorId
        - degradedReason
      properties:
        commandId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - executed_degraded
        result:
          $ref: '#/components/schemas/CommandResult'
        executionMs:
          type: integer
          minimum: 0
        initiatorId:
          type: string
          format: uuid
        degradedReason:
          type: string
          enum:
            - ai_unavailable
            - ai_timeout
            - ai_low_confidence
          description: Why degraded mode was used
        fallbackStrategy:
          type: string
          description: Description of fallback applied
    CommandResult:
      type: object
      properties:
        taskId:
          type: string
          format: uuid
          description: ID of created/affected task
        assigneeId:
          type: string
          format: uuid
          description: ID of assigned user
        decisionConfidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Decision confidence (1.0 for rule-based in Stage 1)
    CommandErrorResponse:
      type: object
      required:
        - correlationId
        - errorCode
        - message
      properties:
        correlationId:
          type: string
          format: uuid
        errorCode:
          type: string
          description: Machine-readable error code (see ErrorCode enum)
        message:
          type: string
          description: Human-readable error message
        validationErrors:
          type: array
          description: Present when errorCode is SCHEMA_INVALID
          items:
            $ref: '#/components/schemas/ValidationError'
        violations:
          type: array
          description: Present when errorCode is BUSINESS_RULE_VIOLATION
          items:
            $ref: '#/components/schemas/BusinessViolation'
    ValidationError:
      type: object
      required:
        - path
        - code
        - message
      properties:
        path:
          type: string
          description: JSON path to invalid field (e.g., $.payload.deadline)
        code:
          type: string
          description: Validation error code
        message:
          type: string
          description: Human-readable validation message
    BusinessViolation:
      type: object
      required:
        - rule
        - message
      properties:
        rule:
          type: string
          enum:
            - ASSIGNEE_MUST_BE_MEMBER
            - ZONE_MUST_EXIST
            - DEADLINE_MUST_BE_FUTURE
            - INITIATOR_MUST_HAVE_PERMISSION
            - TASK_ALREADY_COMPLETED
          description: Business rule that was violated
        message:
          type: string
          description: Human-readable explanation
    AuthErrorResponse:
      type: object
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - AUTH_TOKEN_MISSING
            - AUTH_TOKEN_INVALID
            - AUTH_TOKEN_EXPIRED
        message:
          type: string

    # Domain entities
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
          description: Keycloak subject claim
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        households:
          type: array
          items:
            $ref: '#/components/schemas/HouseholdSummary'
        createdAt:
          type: string
          format: date-time
    HouseholdSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [admin, member]
...